added second heuristic and count of num wins to climbing eval function

+++ src/student_player/mytools/Heuristic.java
@@ -4,5 +4,5 @@ import hus.HusBoardState;
 
 public interface Heuristic {
 	
-	public int evaluate(HusBoardState input, boolean isMyTurn);
+	public double evaluate(HusBoardState input, boolean isMyTurn);
 }
